Added support for custom on-the-fly doxy literals
The `# @command` still works as expected (where the command is a known
parameter), however, to be more flexible, you can add anything you want
in the `command` space if you prefix it with a `!`.  Only the first word
is accepted.  In the Pod, you'll see the command separated by a `:`.
Now, you can do something like:

`# @!EXAMPLE `trim(' test_str    ') => 'test_str'`

which results in something like:

EXAMPLE: `trim(' test_str    ') => 'test_str'`

or

`# @!XS This function is implemented in the MODULE.xs file`
which results in something like:

XS: This function is implemented in the MODULE.xs file